fixed compiler-warnings
git-svn-id: https://lmms.svn.sf.net/svnroot/lmms/trunk/lmms@1368 0778d3d1-df1d-0410-868b-ea421aaaa00d

+++ include/effect_lib.h
@@ -451,7 +451,7 @@ namespace effectLib
 			m_wideCoeff = _wideCoeff;
 		}
 		
-		int getWideCoeff()
+		float getWideCoeff()
 		{
 			return m_wideCoeff;
 		}
